task: Fix samples to pass tests
