+++ Sources/GoodReactor/NSPointerArrayExtension.swift
